requirements.txt
PyYAML
httplib2
oauth2
simplejson
xmpppy
